Resolves data circular dependency.
