+++ assets/js/util/helpers.js
@@ -56,3 +56,31 @@ export const createAddToFilter = ( NewComponent ) => {
 		};
 	};
 };
+
+/**
+ * Replace a filtered component with the passed component and merge their props.
+ *
+ * Components wrapped in the 'withFilters' higher order component have a filter applied to them (wp.hooks.applyFilters).
+ * This helper is used to replace (or "Fill") a filtered component with a passed component. To use, pass as the third
+ * argument to an addFilter call, eg:
+ *
+ * 	addFilter( `googlesitekit.ModuleSettingsDetails-${slug}`,
+ * 		'googlesitekit.AdSenseModuleSettingsDetails',
+ * 		fillFilterWithComponent( AdSenseSettings, {
+ * 			onSettingsPage: true,
+ * 		} ) );
+ *
+ * @param {WPElement} NewComponent The component to render in place of the filtered component.
+ * @param {Object}    newProps     The props to pass down to the new component.
+ *
+ * @return {WPElement} React Component after overriding filtered component with NewComponent.
+ */
+export const fillFilterWithComponent = ( NewComponent, newProps ) => {
+	return ( OriginalComponent ) => {
+		return function InnerComponent( props ) {
+			return (
+				<NewComponent { ...props } { ...newProps } OriginalComponent={ OriginalComponent } />
+			);
+		};
+	};
+};
